Fixed compiler error (gcc).
Problem (MSVC9 compiler does not complain) was in assigning
iterator value to the non constant reference which is not allowed
for std::set. std::set iterators are supposed to be always constant
even if is is not explicitly mentioned. Now I just use the temporary
Filter object, update it and than replace in the container.