Fixed filtering bug.
Fixed filtering bug for cases, when there are more than one
filtering rule defined with the same name. The problem was in
using only one rule with the given name and ignoring the rest
ones because QMultiMap's find function returns the last added
item with the given name. With the given patch we iterate over
all rules that have the given name until the match will be found.