Big refactor of Database class and the whole application concept.
Get rid of the idea of multiple database connections and multiple
database files. The plan is storing everything in the single data
base file that will contain multiple tables with user posts. The
SQL will be completely hidden from the user and he/she don't need
to care about data storing details at all.
Each table in the database will be represented as a saparate folder
for grouping the posts and each filter will store the information
where (in which table) store its data.
Tables can be added/removed etc.
This commit is the start of the global refactoring and code update.