Implemented new function that will emulate LJ login command.
This function validates user's password and get base information
needed for client.
Login to the server, while announcing your client version. The server
returns with whether the password is good or not, the user's name, an
optional message to be displayed to the user, the list of the user's
friend groups, and other things.