Change events properties storing conception.
Store all properties of an event as a string in the one single table
field. This will allow having only one field taked regardless of the
number of properties defined for an event.
To handle properties we concatinate property names and values into
one string and store it in the database.
There are several helper functions added that return event properties.